requirements.txt
pytesseract
Pillow
pandas
